guard role

+++ src/item/items.service.ts
@@ -1,9 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
 import { ItemStatus } from './entity/item-status.enum';
 import { CreateItemDto } from './dio/create-item.dio';
 import { Item } from './entity/item.entity';
 import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/user/entity/user.entity';
 
 
 
@@ -26,7 +27,7 @@ export class ItemsService {
     return item;
   }
 
-  async create(createItemDto: CreateItemDto): Promise<Item> {
+  async create(createItemDto: CreateItemDto, user: User): Promise<Item> {
     const { name, price, description } = createItemDto;
 
     const item = this.itemsRepository.create({
@@ -36,17 +37,25 @@ export class ItemsService {
       status: ItemStatus.ON_SALE,
       createdAt: new Date().toISOString(),
       updatedAt: new Date().toISOString(),
+      user: user,
     });
     return this.itemsRepository.save(item);
   }
 
-  async updateStatus(id: string): Promise<Item> {
+  async updateStatus(id: string, user: User): Promise<Item> {
     const item = await this.findById(id);
+    if (item.userId === user.id) {
+      throw new BadRequestException(`自身の商品は購入できません`);
+    }
     item.status = ItemStatus.SOLD_OUT;
     return this.itemsRepository.save(item);
   }
 
-  async delete(id: string): Promise<void> {
+  async delete(id: string, user: User): Promise<void> {
+    const item = await this.findById(id);
+    if (item.userId !== user.id) {
+      throw new BadRequestException(`自身の商品以外は削除できません`);
+    }
     this.itemsRepository.delete(id);
   }
 }
